Add custom user model, remove old user model, upgrade Python and Django
