+++ costcentre/templates/costcentre/admin/change_permission_form.html
@@ -44,6 +44,7 @@
                     <th></th>
                     <th scope="col" class="sortable column-user_email">First name</th>
                     <th scope="col" class="sortable column-user_first_name">Last name</th>
+                    <th scope="col" class="sortable column-user_email">Email</th>
                 </tr>
             </thead>
             <tbody>
@@ -54,6 +55,7 @@
                         </td>
                         <td>{{ user.first_name }}</td>
                         <td>{{ user.last_name }}</td>
+                        <td>{{ user.email }}</td>
                     </tr>
                 {% endfor %}
                 <tr>
