+++ docker-compose.yml
@@ -50,14 +50,3 @@ services:
       - HUB_PORT=4444
     ports:
       - "5900:5900"
- firefox:
-    image: selenium/node-firefox-debug:3.141.59-vanadium
-    volumes:
-      - /dev/shm:/dev/shm
-    depends_on:
-      - selenium-hub
-    environment:
-      - HUB_HOST=selenium-hub
-      - HUB_PORT=4444
-    ports:
-      - "5901:5900"
