+++ docker/Dockerfile.dev
@@ -1,4 +1,4 @@
-FROM python:3.6
+FROM python:3.9.0
 ADD . /app
 WORKDIR /app
