+++ forecast/admin.py
@@ -23,14 +23,6 @@ from forecast.models import (
 User = get_user_model()
 
 
-# def get_name(self):
-#     return '{} {}'.format(self.first_name, self.last_name)
-#
-#
-# #  TODO - swap user model to modern version, monkey patch for now
-# User.add_to_class("__str__", get_name)
-
-
 class MonthlyFigureAdmin(AdminImportExport, AdminReadOnly, SimpleHistoryAdmin):
     @property
     def import_info(self):
