+++ gifthospitality/filters.py
@@ -1,10 +1,10 @@
-from bootstrap_datepicker_plus import DatePickerInput
-
 from django_filters import (
     DateFilter,
     ModelChoiceFilter,
     NumberFilter,
 )
+
+from django.forms import DateInput
 
 from core.filters import MyFilterSet
 
@@ -85,13 +85,7 @@ class GiftHospitalityFilter(MyFilterSet):
         field_name="entered_date_stamp",
         label="Date Entered From:",
         lookup_expr="gte",
-        widget=DatePickerInput(
-            options={
-                "format": "DD/MM/YYYY",  # moment date-time format
-                "showClose": True,
-                "showClear": True,
-                "showTodayButton": True,
-            },
+        widget=DateInput(
             attrs={
                 "placeholder": "DD/MM/YYYY",
             },
@@ -102,13 +96,7 @@ class GiftHospitalityFilter(MyFilterSet):
         field_name="entered_date_stamp",
         label="To:",
         lookup_expr="lte",
-        widget=DatePickerInput(
-            options={
-                "format": "DD/MM/YYYY",  # moment date-time format
-                "showClose": True,
-                "showClear": True,
-                "showTodayButton": True,
-            },
+        widget=DateInput(
             attrs={
                 "placeholder": "DD/MM/YYYY",
             },
