+++ gifthospitality/templates/gifthospitality/giftandhospitality_added.html
@@ -1,6 +1,4 @@
 {% extends "standard_content.html" %}
-{% load bootstrap4 %}
-{% bootstrap_css %}
 
 {% block title %}Gift &amp; Hospitality - Added Entry{% endblock %}
