+++ gifthospitality/templates/gifthospitality/giftandhospitality_form.html
@@ -1,7 +1,4 @@
 {% extends "standard_content.html" %}
-{% load bootstrap4 %}
-{% bootstrap_css %}
-{% bootstrap_javascript jquery='full' %}
 {% block scripts %}
 <script>
 var body = document.getElementsByTagName("body")[0]
