+++ gifthospitality/templates/gifthospitality/gifthospitality_filter.html
@@ -3,10 +3,7 @@
 {% load render_table from django_tables2 %}
 {% load querystring from django_tables2 %}
 
-{% load bootstrap4 %}
 {% load gifthospitality_permissions %}
-{% bootstrap_css %}
-{% bootstrap_javascript jquery='full' %}
 
 {% block title %}Gift &amp; Hospitality - Search Records{% endblock %}
