+++ requirements.in/base.in
@@ -8,39 +8,36 @@ botocore>=1.12.42
 celery==4.4.4
 certifi==2018.8.24
 chardet==3.0.4
-custom-usermodel @ git+https://github.com/uktrade/django-staff-sso-usermodel.git@c8d0c0ed17c2b39464b98b3cf93660a4db7d67a1
 decorator==4.2.1
 diff-match-patch==20121119
 dj-database-url==0.5.0
-Django==2.2.13
-django-admin-list-filter-dropdown==1.0.1
-django-adv-cache-tag==1.1.3
-django-appconf==1.0.4
-django-autocomplete-light==3.3.2
-django-axes==5.3.2
-django-bootstrap-breadcrumbs==0.9.2
-django-bootstrap-datepicker-plus==3.0.5
-django-bootstrap4==0.0.7
-django-compressor==2.4
-django-environ==0.4.5
-django-extensions==2.1.2
-django-filter==2.0.0
-django-guardian==2.2.0
-django-import-export==1.1.0
-django-ipware==2.1.0
-django-log-formatter-ecs==0.0.3
-django-pivot==1.7
-django-redis==4.12.1
-django-reversion==3.0.7
-django-sass-processor==0.8
-django-settings-export==1.2.1
-django-simple-history==2.10.0
-django-staff-sso-client==2.2.1
-django-storages>=1.7.1
-django-tables2==2.1.1
-django-tablib==3.2
-django-webpack-loader==0.7.0
-djangorestframework==3.11.0
+Django==3.0.3
+django-admin-list-filter-dropdown
+django-adv-cache-tag
+django-appconf
+django-autocomplete-light
+django-axes
+django-compressor
+django-environ
+django-extensions
+django-filter
+django-guardian
+django-import-export
+django-ipware
+django-log-formatter-ecs
+django-pivot
+django-redis
+django-reversion
+django-sass-processor
+django-settings-export
+django-simple-history
+django-staff-sso-client
+django-storages
+django-tables2
+django-tablib
+django-webpack-loader
+django-bootstrap-breadcrumbs
+djangorestframework
 docutils==0.14
 elastic-apm==5.6.0
 et-xmlfile==1.0.1
@@ -50,8 +47,6 @@ future==0.18.2
 gunicorn==19.9.0
 idna==2.7
 importlib-metadata==1.6.0
-ipython==6.2.1
-ipython-genutils==0.2.0
 jdcal==1.3
 jedi==0.11.1
 jmespath==0.9.3
@@ -69,8 +64,7 @@ pexpect==4.4.0
 pickleshare==0.7.4
 pluggy>=0.7.1
 prompt-toolkit==1.0.15
-psycopg2==2.7.4
-psycopg2-binary==2.7.4
+psycopg2>=2.8.5,<3
 ptyprocess==0.5.2
 public==1.0.3
 py==1.6.0
@@ -90,7 +84,7 @@ redis==3.5.3
 requests==2.21.0
 requests-oauthlib==1.0.0
 rjsmin==1.1.0
-s3chunkuploader==0.12
+s3chunkuploader
 s3transfer>=0.1.13
 sentry-sdk==0.10.2
 simplegeneric==0.8.1
@@ -98,12 +92,11 @@ six>=1.11.0
 sqlparse==0.2.4
 tablib==0.12.1
 text-unidecode>=1.2
-traitlets==4.3.2
 unicodecsv==0.14.1
 urllib3==1.24.2
 vine==1.3.0
 wcwidth==0.1.7
-whitenoise==4.1
+whitenoise==5.2
 xlrd==1.1.0
 xlwt==1.3.0
 zipp==3.1.0
