+++ runtime.txt
@@ -1,1 +1,1 @@
-python-3.6.12
+python-3.9.0
