+++ user/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class UserConfig(AppConfig):
+    name = "user"
+    verbose_name = "User config"
