+++ user/models.py
@@ -1,0 +1,6 @@
+from django.contrib.auth.models import AbstractUser
+
+
+class User(AbstractUser):
+    def __str__(self):
+        return f"{self.first_name} {self.last_name}"
